test(CurrencyLabel): remove using enzyme (#3158)
